+++ src/images/icon/icons.js
@@ -1,5 +1,5 @@
 import openEye from './openEye.svg';
-import eye from './eye.svg'
+import eye from './eye.svg';
 import closedEye from './closedEye.svg';
 import eyeOff from './eye-off.svg';
 import logoNide from './logoNide.svg';
@@ -9,6 +9,7 @@ import nideMotors from './nideMotors.svg';
 import logoUsmotors from './logoUsmotors.svg';
 import smallArrowRight from './smallArrowRight.svg';
 import plus from './plus.svg';
+import whitePlus from './whitePlus.svg';
 import downArrow from './downArrow.svg';
 import dashboardIcon from './dashboardIcon.svg';
 import alertCircle from './alertCircle.svg';
@@ -31,6 +32,7 @@ export default {
   logoUsmotors,
   smallArrowRight,
   plus,
+  whitePlus,
   downArrow,
   dashboardIcon,
   alertCircle,
